requirements.txt
grpcio
grpcio-tools
numpy
Pillow
scipy